# Make GPU kernel for argsort use int32