--- a/src/operator/tensor/init_op.h
+++ b/src/operator/tensor/init_op.h
@@ -637,8 +637,8 @@ void EyeFill(const nnvm::NodeAttrs& attrs,
 
 
 struct range_fwd {
-  template<typename DType>
-  MSHADOW_XINLINE static void Map(index_t i, index_t repeat, DType start, DType step,
+  template<typename DType, typename IDXType = index_t>
+  MSHADOW_XINLINE static void Map(index_t i, IDXType repeat, DType start, DType step,
                                   int req, DType* out) {
     KERNEL_ASSIGN(out[i], req, start + (i/repeat) * step);
   }